move transform_tree out of init file
